refactor value layout for easier additional frameworks
